test: Move replicated legacy tests to separate file
These don't need to pollute the primary file.
Otherwise, future tests may feel compelled to duplicate for legacy and
project variants.
